Better place for test helper
