+++ compiler/noirc_evaluator/src/ssa.rs
@@ -47,7 +47,6 @@ pub mod ir;
 mod opt;
 mod parser;
 pub mod ssa_gen;
-mod tests;
 
 pub struct SsaEvaluatorOptions {
     /// Emit debug information for the intermediate SSA IR
