+++ compiler/noirc_evaluator/src/ssa/opt/mod.rs
@@ -3,6 +3,7 @@
 //! Each pass is generally expected to mutate the SSA IR into a gradually
 //! simpler form until the IR only has a single function remaining with 1 block within it.
 //! Generally, these passes are also expected to minimize the final amount of instructions.
+
 mod array_set;
 mod as_slice_length;
 mod assert_constant;
@@ -21,3 +22,17 @@ mod resolve_is_unconstrained;
 mod runtime_separation;
 mod simplify_cfg;
 mod unrolling;
+
+#[cfg(test)]
+pub(crate) fn assert_ssa_equals(mut ssa: super::Ssa, expected: &str) {
+    ssa.normalize_ids();
+
+    let ssa = ssa.to_string();
+    let ssa = ssa.trim();
+    let expected = expected.trim();
+
+    if ssa != expected {
+        println!("Expected:\n~~~\n{}\n~~~\nGot:\n~~~\n{}\n~~~", expected, ssa);
+        similar_asserts::assert_eq!(expected, ssa);
+    }
+}
